test: submitOrder 테스트를 order.test.js에서 order-submit.test.js로 분리 (#186)
